==> 1.extractSeq.R <==
#The ./Dataset directory contains the sequences used in the study.
# Worm from MethSMRT, ref:ce10, 6mA, identityQv:30, CD-hit c 0.6
# Human from Mol cell 2018, ref:hg38, 6mA, identityQv:30, CD-hit c 0.6
# Rice from Nat Plant 2018, ref:MH63, 6mA, identityQv:30, CD-hit c 0.6. num:880
# Fly from MethSMRT, ref:xx, 6mA, identityQv:30, CD-hit c 0.6

#Rice neg from iPred-6mA paper
#The left neg samples will be extracted from the genomes. Details are documented in the 2_Negative_samples.R script
